Configuration gate for feature * Put source and thumbnail deletion functionality behind configuration * Updated tests with new configuration parameter

+++ server/szurubooru/func/posts.py
@@ -400,8 +400,9 @@ def _before_post_delete(
         _mapper: Any, _connection: Any, post: model.Post) -> None:
     if post.post_id:
         image_hash.delete_image(str(post.post_id))
-        files.delete(get_post_content_path(post))
-        files.delete(get_post_thumbnail_path(post))
+        if config.config['delete_source_files']:
+            files.delete(get_post_content_path(post))
+            files.delete(get_post_thumbnail_path(post))
 
 
 def _sync_post_content(post: model.Post) -> None:
